+++ src/containerapp/azext_containerapp/_params.py
@@ -295,6 +295,11 @@ def load_arguments(self, _):
         c.argument('max_executions', type=int, help="Maximum number of job executions that are created for a trigger", default=100)
         c.argument('polling_interval', type=int, help="Interval to check each event source in seconds.", default=30)
 
+    # scale
+    with self.argument_context('containerapp', arg_group='Scale') as c:
+        c.argument('scale_rule_identity', options_list=['--scale-rule-identity', '--sri'], 
+                   help='Resource ID of a managed identity to authenticate with Azure scaler resource(storage account/eventhub or else), or System to use a system-assigned identity.', is_preview=True)
+
     with self.argument_context('containerapp job', arg_group='Scale') as c:
         c.argument('min_executions', type=int, help="Minimum number of job executions to run per polling interval.")
         c.argument('max_executions', type=int, help="Maximum number of job executions to run per polling interval.")